csvAI EvaDB/requirements.txt
langchain
evadb
sentence-transformers
pandas
numpy
openai
tiktoken
faiss-cpu
python-dotenv